% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{build_doors}
\alias{build_doors}
\title{builds doors for the new Monty Hall Problem game with 5 doors- 3 goats and 2 cars}
\usage{
build_doors(n = 5)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
returns a numeric vector of door numbers (1:n), not goat/car labels
}
\description{
build_doors() generates two more doors n =5
3 doors with goats behind them, and two doors with cars.
}
\details{
The game setup.
}
\examples{
  build_doors()

}
