% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{change doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{'stay'}{is a logical value that indicates whether the
contestant stayed with their original pick (TRUE) which is the
default or switches to the other unopened door (`FALSE).}

\item{'opened.door'}{A length 1 numeric vector indicating the
door opened by the host, as returned by 'open_goat_door()`.
also located in the ELSE statement as something to exclude
(!doors \%in\% c(a.pick, opened.door)) — so it must be the door
that's no longer a valid choice because the host already opened it.}

\item{'a.pick'}{A length 1 numeric vector indicating the
contestant's initial door pick, as returned by
'select_door()`.}
}
\value{
The function returns a length 1 numeric vector
indicating the contestant's final door pick (1, 2, or 3).
}
\description{
`change_doors()" determines the contestant's final door
selection, based on whether they stay with their original
choice or switch to the remaining unopened door.
}
\details{
After the host opens a goat door, the contestant may
either stay with their original selection or switch to
the other unopened door. This function returns the final
decision.
}
\examples{
opened.door <- open_goat_door( this.game, my.initial.pick )
change_door( stay=T, 
            opened.door=opened.door, 
            a.pick=my.initial.pick )
change_door( stay=F, 
           opened.door=opened.door, 
            a.pick=my.initial.pick )
my.final.pick <- change_door( stay=F, 
                             opened.door=opened.door, 
                             a.pick=my.initial.pick )
this.game
my.initial.pick
my.final.pick 

}
