% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door1}
\alias{change_door1}
\title{Stay with initial pick or switch to another door}
\usage{
change_door1(stay = T, opened.doors, a.pick)
}
\arguments{
\item{opened.doors}{A length 2 numeric vector, as returned by
open_doors1(), indicating the two doors that were opened by the
host.}

\item{a.pick}{A length 1 numeric vector, as returned by
select_door1(), indicating the contestant's original door pick.}

\item{(stay)}{a logical value indicating whether the contestant
stays with their original pick (TRUE, the default) or switches
to a different door (FALSE).}
}
\value{
The function returns a length 1 numeric vector, a number
between 1 and 5 representing the contestant's final door
selection.
}
\description{
change_door1() determines the contestant's final door choice,
based on whether they choose to stay with their original pick or
switch to another unopened door.
}
\details{
After the host opens two doors (one car, one goat), the
contestant is given the option to stay with their original
selection or switch to one of the other unopened,
doors. This function implements both statements  if stay=TRUE,
the contestant's final pick is unchanged; if stay=FALSE, a new
door is randomly selected from the doors that were neither
opened by the host or originally picked by the contestant.
}
\examples{
this.game <- create_game1()
my.pick <- select_door1()
opened <- open_doors1( this.game, my.pick )
change_door1( stay=TRUE, opened.doors=opened, a.pick=my.pick )
change_door1( stay=FALSE, opened.doors=opened, a.pick=my.pick )

}
