% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game1}
\alias{create_game1}
\title{Create a new Monty Hall Problem game with 5 doors- 3 goats and 2 cars}
\usage{
create_game1()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 5 character vector
indicating the positions of goats and the car.
}
\description{
create_game1() generates a new game that consists of five doors
with goats behind them, and two with cars.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens two door to reveal a goat and car, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  create_game1()

}
