% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game2}
\alias{create_game2}
\title{Create a Dynamic World of the Monty Hall Problem game}
\usage{
create_game2(num.goats, num.cars)
}
\arguments{
\item{num.goats}{Integer the number of "goat" doors to include
in the game.}

\item{num.cars}{Integer rhe number of "car" doors to include in
the game.}
}
\value{
The function returns a character vector of length
num.goats + num.cars containing a random arrangement of
"goat" and "car" values.
}
\description{
create_game2() generates a new game consisting of a mix of
"goat" and "car" doors, based on the specified number of each.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are doors for a contestant to choose
from, some of which have a car behind them and some have goats.
This version generalizes the classic 3-door setup, allowing any
number of goat doors and car doors to be specified, and the total
number of doors is the sum of the two. The doors are shuffled
randomly so the position of the car(s) and goat(s) is not
predictable.
}
\examples{
create_game2( num.goats=2, num.cars=1 )

}
