% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{determines if contestant won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{game}{is a length of 3 character vector and returned as 'create_game' indicating
the locations of goats and the car.}

\item{'final.pick'}{is a length of 1 numeric vector indicating the contestants final
pick and resturned as 'change_door()'}
}
\value{
The function returns a character string, either
"WIN" or "LOSE" based on whether the final pick was the car or a goat.
}
\description{
determine_winner() checks whether the contestant's final selection
has the car
}
\details{
compares the final door pick with the underlying game to determine whether
the outcome is the winner 'car' or loser 'goat'.
}
\examples{
this.game
my.initial.pick
my.final.pick <- change_door( stay=T, 
                             opened.door=opened.door, 
                             a.pick=my.initial.pick )
determine_winner( final.pick=my.final.pick, 
                 game=this.game )
my.final.pick <- change_door( stay=F, 
                             opened.door=opened.door, 
                             a.pick=my.initial.pick )
determine_winner( final.pick=my.final.pick, 
                game=this.game )
}
