% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner1}
\alias{determine_winner1}
\title{Determine if the contestant has won or lost}
\usage{
determine_winner1(final.pick, game)
}
\arguments{
\item{final.pick}{A length 1 numeric vector, as returned by
change_door1()`, indicating the contestant's final door
selection.}

\item{game}{A length 5 character vector, as returned by
create_game1(), indicating what is behind each door.}
}
\value{
The function returns a character string, either "WIN" if
the final pick was a car, or "LOSE" if the final pick was a
goat.
}
\description{
determine_winner1() checks the contestant's final door pick
against the game setup to determine whether they won a car or
lost with a goat.
}
\details{
This function compares what is actually behind the contestant's
final chosen door to determine the outcome of the game, after the
contestant has decided whether to stay with their original pick
or switch doors.
}
\examples{
this.game <- create_game1()
my.pick <- select_door1()
opened <- open_doors1( this.game, my.pick )
final <- change_door1( stay=TRUE, opened.doors=opened, a.pick=my.pick )
determine_winner1( final, this.game )

}
