% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner2}
\alias{determine_winner2}
\title{Determine if the contestant has won.}
\usage{
determine_winner2(final.pick, game)
}
\arguments{
\item{final.pick}{Integer the contestant's final door pick, as
returned by change_door2().}

\item{game}{A character vector representing the game setup, as
returned by create_game2(), containing "goat" and "car" values.}
}
\value{
The function returns a length 1 character string, either
"WIN" if the final pick is a car, or "LOSE" if the final pick is
a goat.
}
\description{
determine_winner2() determines whether the contestant's final
door pick results in a win or a loss, based on whether the door
hides a car or a goat.
}
\details{
Once the contestant has made their final decision, this function
checks what is behind their chosen door and reports the outcome
of the game.
}
\examples{
this.game <- create_game2( num.goats=2, num.cars=1 )
determine_winner2( final.pick=3, game=this.game )

}
