% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-100doors.R
\name{final_contestant_choice}
\alias{final_contestant_choice}
\title{Determine whether contestant decides to take the final envelope or keeps their original
choice.}
\usage{
final_contestant_choice(
  envelopes,
  contestant_choice,
  host_choice,
  switch_choice
)
}
\arguments{
\item{envelopes}{Character vector}

\item{contestant_choice}{Numeric integer}

\item{host_choice}{Numeric vector, 98 numbers}

\item{switch_choice}{Character, string to determine whether to SWITCH or KEEP}
}
\value{
The function returns a character string of newsprint or $1000
}
\description{
\code{final_contestant_choice()} filters out all envelope options to remove the contestant
and host's combined 99 selections and leaves one envelope. Also allows contestant
to decide whether to keep their original envelope or switch.
}
\details{
This function is the third step in the game and ultimately determines whether the
contestant wins or loses
}
\examples{
final_contestant_choice(1:100, 13, 1:98, "SWITCH")
}
