% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-100doors.R
\name{monty_hall_100doors}
\alias{monty_hall_100doors}
\title{Run through game and store outcome values}
\usage{
monty_hall_100doors(switch_choice)
}
\arguments{
\item{switch_choice}{Character string, "KEEP" or "SWITCH"}
}
\value{
List with labels for stored values
}
\description{
\code{monty_hall_100doors()} plays the entire game and returns a list of game outcomes
}
\details{
The function provides an output of all contents of the 100 envelopes, the contestant's
original choice, the host's choice of 98 envelopes, the contestant's decision to switch,
the final choice, and the final game outcome
}
\examples{
monty_hall_100doors("SWITCH")
monty_hall_100doors("KEEP")
}
