% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_doors1}
\alias{open_doors1}
\title{Open doors to reveal a car and a goat}
\usage{
open_doors1(game, a.pick)
}
\arguments{
\item{game}{A length 5 character vector, as returned by
create_game1(), indicating what is behind each door.}

\item{a.pick}{A length 1 numeric vector, as returned by
select_door1(), indicating the door the contestant initially
selected.}
}
\value{
The function returns a length 2 numeric vector containing
the numbers of the two doors that were opened: one with a car,
one with a goat.
}
\description{
open_doors1() simulates the host opening two of the remaining,
un selected doors one always revealing one car and the other one goat. This is after
the contestant has made their initial pick.
}
\details{
Depending on whether the contestant's initial pick was a car or a
goat, this function selects two doors from among the doors not
picked by the contestant. Meaning one door with a car behind it, and one
door with a goat behind it. This is suppose to be the host's role in the
Monty Hall game, who always reveals a combination of doors
without revealing the contestant's own door.
}
\examples{
this.game <- create_game1()
my.pick <- select_door1()
open_doors1( this.game, my.pick )

}
