% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{open a goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{obtains a length of a 3 character vector and returned by 'create_game()'
indicating the locations of the goats and the car}

\item{a.pick}{obtains a length of 1 numeric vector indicating the contestants original door pick
and returned as 'select_door()'}
}
\value{
the function returns a length 1 numeric vector indicating the door numbers (1-3)
opened by the host which will always reveal a goat
}
\description{
open_goat_door () host opens one of the two unselected doors that has a goat
}
\details{
if the contestant originally picks the door with the car the host will randomly open
1/2 remaining doors. If the contestant initially picked the door with the goat
the host opens the other door with the goat making sure to not reveal the door with the car
}
\examples{
this.game <- create_game()
this.game
my.initial.pick <- select_door()
my.initial.pick
open_goat_door(this.game, my.initial.pick)

}
