% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-100doors.R
\name{select_contestant_choice}
\alias{select_contestant_choice}
\title{Simulate contestant's selection of first envelope}
\usage{
select_contestant_choice()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a integer between 1 and 100
}
\description{
\code{select_contestant_choice()} utilizes 'sample' to select a number between 1-100 that
corresponds to the number of envelopes
}
\details{
This function is the first step after starting the game and represents a 1/100 chance of
selecting the $1000 envelope.
}
\examples{
select_contestant_choice()
}
