% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door1}
\alias{select_door1}
\title{Select a door}
\usage{
select_door1()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 numeric vector, a number
between 1 and 5 representing the door the contestant picked.
}
\description{
select_door1() randomly selects one of the five doors as the
contestant's initial pick.
}
\details{
This simulates the contestant's first move in the game to choosing
one door out of five, before any doors have been opened by the
host. The selection is made completely at random.
}
\examples{
select_door1()

}
