% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door2}
\alias{select_door2}
\title{Select a door.}
\usage{
select_door2(num.doors)
}
\arguments{
\item{num.doors}{Integer the total number of doors in the game.}
}
\value{
The function returns a length 1 integer between 1 and
num.doors , indicating the contestant's initial door pick.
}
\description{
select_door2() randomly selects one door out of the total
number of doors available, simulating the contestant's initial pick.
}
\details{
In the Monty Hall game, the contestant's first step is to pick
one of the doors without knowing what is behind any of them.
This function simulates that random initial choice.
}
\examples{
 select_door2( num.doors=3 )

}
