% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-100doors.R
\name{select_host_choice}
\alias{select_host_choice}
\title{Simulate host's selection of 98 envelopes containing newsprint}
\usage{
select_host_choice(envelopes, contestant_choice)
}
\arguments{
\item{envelopes}{Character vector}

\item{contestant_choice}{Numeric integer}
}
\value{
The function returns a numeric vector
}
\description{
\code{select_host_choice()} utilizes the R functions which and sample to randomly
select 98 envelopes that have newsprint in them.
}
\details{
This function is the second step in the game and eliminates almost all of the remaining
envelopes available for the contestant. Also guarantees that only newsprint envelopes
are eliminated.
}
\examples{
select_host_choice(1:100, 13)
}
