% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-100doors.R
\name{start_game}
\alias{start_game}
\title{Create a new Monty Hall Problem game.}
\usage{
start_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 3 character vector
indicating the positions of goats and the car.
}
\description{
\code{start_game()} generates a new game that consists of 100 envelopes
with 99 filled with newsprint and one with a $1000 dollars.
}
\details{
The game setup replicates an alternate version of the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  start_game()

}
